requirements.txt
aiohttp==3.9.1
click==8.1.7
fvalues==0.0.4
google_api_python_client==2.113.0
google_auth_oauthlib==1.2.0
gorilla==0.4.0
graphviz==0.20.1
ipdb==0.13.13
ipython==8.20.0
langchain==0.0.306
matplotlib==3.8.2
merge_args==0.1.5
nltk==3.8.1
Pillow==10.2.0
protobuf==4.25.2
pydantic==1.10.13
pymongo==4.6.1
python_dateutil==2.8.2
PyYAML==6.0.1
Requests==2.28.1
rich==13.7.0
rouge_score==0.1.2
setuptools==65.5.0
tenacity==8.2.3
tiktoken==0.5.2
torch==2.1.2
tyro==0.5.6
python-box
openai==0.27.10
pyowm
open-clip-torch
sentence-transformers
chromadb==0.4.8
